+++ MundoVoxel.Client/Juego/teclas.cs
@@ -6,6 +6,7 @@ public static class Teclas
     public const int W = 0x57, A = 0x41, S = 0x53, D = 0x44;
     public const int Espacio = 0x20, F = 0x46, T = 0x54, Escape = 0x1B, R = 0x52, Mayus = 0x10;
     public const int Num1 = 0x31, Num9 = 0x39, E = 0x45, Q = 0x51, U = 0x55;
+    public const int G = 0x47; // modo espectador
 
     // Códigos virtuales para los controles táctiles (Android)
     public const int Volar = 0x100, Chat = 0x101, Pausa = 0x102,