+++ MundoVoxel.Core/bloques.cs
@@ -33,6 +33,7 @@ public static class Bloques
     public const ushort Antorcha = 27;
     public const ushort Cobre = 28;
     public const ushort Cofre = 29;
+    public const ushort Lava = 30;
 
     public sealed record InfoBloque(string ClaveLang, bool Solido, bool Transparente, bool Liquido);
 
@@ -69,13 +70,14 @@ public static class Bloques
         new("bloque.antorcha", false, true,  false), // 27
         new("bloque.cobre",    true,  false, false), // 28
         new("bloque.cofre",    true,  false, false), // 29
+        new("bloque.lava",     false, true,  true ), // 30
     };
 
     public static bool EsSolido(ushort b) => b < Info.Length && Info[b].Solido;
     public static bool EsTransparente(ushort b) => b < Info.Length && Info[b].Transparente;
     public static bool EsLiquido(ushort b) => b < Info.Length && Info[b].Liquido;
-    public static bool EsColocable(ushort b) => b > Aire && b != Lecho && b != Agua && !EsCultivo(b) && b != TierraLabrada && b != Planton;
-    public static bool EsRompible(ushort b) => b != Aire && b != Lecho;
+    public static bool EsColocable(ushort b) => b > Aire && b != Lecho && b != Agua && b != Lava && !EsCultivo(b) && b != TierraLabrada && b != Planton;
+    public static bool EsRompible(ushort b) => b != Aire && b != Lecho && b != Agua && b != Lava;
     public static bool EsMineral(ushort b) => b == Carbon || b == Hierro || b == Oro || b == Diamante || b == Cobre;
     public static bool EsCultivo(ushort b) => b >= Trigo0 && b <= Trigo3;
     public static bool EsTrigoMaduro(ushort b) => b == Trigo3;